Add files via upload
Senior Level Project where I had the opportunity to be the project manager for a systems programming group assignment in which our team of three successfully produced a modular program that interpreted, sorted, and stored binary data for airline flights into an in-memory database accessible through a command-line text-based user interface. (Unix VM) My contributions were: Indexer.c fileconverter.c and srchindx.c and I assisted in developing all of the remaining code as it all had to work in unison.